+++ changelogs/unreleased/6010-replace-json-mocks.yml
@@ -1,0 +1,4 @@
+description: Remove unused mocks and move JSON mocks to typescript
+issue-nr: 6010
+change-type: patch
+destination-branches: [master]
