+++ src/UI/Components/Diagram/Mocks/index.ts
@@ -1,1 +1,2 @@
-export * from "./Mock";
+export * from "./isntances";
+export * from "./services";
