+++ src/UI/Components/ServiceInstanceForm/Components/RelatedServiceProvider.tsx
@@ -9,7 +9,7 @@ interface Props {
   serviceName: string;
   attributeName: string;
   attributeValue: string | string[] | null;
-  description?: string;
+  description?: string | null;
   isOptional: boolean;
   handleInputChange: (value) => void;
   alreadySelected: string[];
@@ -23,7 +23,7 @@ interface Props {
  * @prop {string} serviceName - The name of the service.
  * @prop {string} attributeName - The name of the attribute.
  * @prop {string | string[] | null} attributeValue - The value of the attribute.
- * @prop {string} description - The description of the attribute.
+ * @prop {string | null} description - The description of the attribute.
  * @prop {boolean} isOptional - Whether the attribute is optional.
  * @prop {(value): void} handleInputChange - The function to handle input changes.
  * @prop {string[]} alreadySelected - The already selected options.
